Add add rector and phpstan:
- add configuration files
- refactored code following rector and phpstan analyses

+++ src/JsonRpc/Exception/InternalErrorException.php
@@ -10,6 +10,6 @@ class InternalErrorException extends JsonRpcException
 {
     public function __construct()
     {
-        parent::__construct(JsonRpcError::InternalError->message(), JsonRpcError::InternalError->code());
+        parent::__construct(JsonRpcError::InternalError->message(), (int)JsonRpcError::InternalError->code());
     }
 }
